pondtv/requirements.txt
PyYAML
psutil
python-mpv
pynput 